daily_farm/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
python-dotenv==1.0.0
alembic==1.13.1
pydantic==2.10.3
pydantic-settings==2.2.1
redis==5.0.1
boto3==1.34.11
httpx==0.26.0
python-magic==0.4.27
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.23.2
emails==0.6
httpx-oauth==0.13.1 